+++ pkg/datastore/redis/keys.go
@@ -1,4 +1,4 @@
-package datastore
+package redis
 
 import ()
