+++ pkg/datastore/redis/storesecret.go
@@ -1,10 +1,9 @@
-package datastore
+package redis
 
 import (
 	"context"
 	"errors"
 	"fmt"
-	"log"
 )
 
 // StoreSecret stores
@@ -29,19 +28,4 @@ func (d *Datastore) StoreSecret(ctx context.Context, id string, secret string) e
 	}
 
 	return nil
-}
-
-// FetchSecret fetches
-func (d *Datastore) FetchSecret(ctx context.Context, id string) (string, error) {
-
-	log.Println("datastore.FetchSecret():", id)
-
-	key := fmt.Sprintf(defaultIdentitySecretKeyFormat, id)
-
-	secret, err := d.client.Get(d.ctx, key).Result()
-	if err != nil {
-		return "", err
-	}
-
-	return secret, nil
 }
